Moved post to a separete file.
