+++ serverfiles/uploads/form_multipart.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Welcome to Webserv</title>
+</head>
+<body>
+    <h1>Upload form!</h1>
+    <form action="/uploads" method="POST" enctype="multipart/form-data">
+    <input type="file" name="file">
+    <input type="submit" value="Upload">
+</form>
+</body>
+</html>
