+++ serverfiles/uploads/form_urlencoded.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>URL Encoded Form Test</title>
+</head>
+<body>
+    <h1>Test application/x-www-form-urlencoded</h1>
+    <form action="/uploads" method="POST">
+        <label for="username">Username:</label>
+        <input type="text" name="username" id="username"><br><br>
+
+        <label for="email">Email:</label>
+        <input type="email" name="email" id="email"><br><br>
+
+        <label for="message">Message:</label><br>
+        <textarea name="message" id="message" rows="5" cols="30"></textarea><br><br>
+
+        <input type="submit" value="Send">
+    </form>
+</body>
+</html>
